Adicionado comentários no código

+++ c++/engine.cpp
@@ -29,29 +29,51 @@ class Engine {
 		 */
 		const static unsigned EXCEPTION_INVALID_UNDER_RANGE = 1;
 
+		/**
+		 * Construtor padrão
+		 */
 		Engine() {
 			this->setYear(DEFAULT_YEAR);
 			this->setPower(ZERO_FLOAT);
 		}
 
+		/**
+		 * Construtor com inicialização externa de variáveis
+		 */
 		Engine(unsigned y, float p) {
 			this->setYear(y);
 			this->setPower(p);
 		}
 
+		/**
+		 * Getters & Setters
+		 */
+
+		/**
+		 * Setter do ano do motor
+		 * @param y ano
+		 */
 		void setYear(unsigned y) {
 
-			if(y < DEFAULT_YEAR)
+			if(y < (unsigned) DEFAULT_YEAR)
 				throw 1;
 
 			this->year = y;
 
 		}
 
+		/**
+		 * Getter do ano do motor
+		 * @return ano
+		 */
 		unsigned getYear() {
 			return this->year;
 		}
 
+		/**
+		 * Setter da potência do motor
+		 * @param p potência em CV
+		 */
 		void setPower(float p) {
 			if(p < ZERO_FLOAT)
 				throw 1;
@@ -59,10 +81,22 @@ class Engine {
 			this->power = p;
 		}
 
+		/**
+		 * Getter da potência do motor
+		 * @return potência
+		 */
 		float getPower() {
 			return this->power;
 		}
 
+		/**
+		 * Métodos de negócio
+		 */
+
+		/**
+		 * Valida se as propriedades são válidas para validar o motor
+		 * @return true se for válido, false se não
+		 */
 		bool isValid() {
 			if(this->year < DEFAULT_YEAR || this->power < ZERO_FLOAT)
 				return false;
